backend/requirements.txt
flask
flask-cors
firebase-admin
google-cloud-vision
google-generativeai
python-dotenv
gunicorn
